requirements.txt
pytest
pytest-cov
pandas
python-pptx